Fix settings types

+++ packages/valist-web/forms/update-project.tsx
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { ApolloCache } from '@apollo/client';
-import { ProjectMeta, Client } from '@valist/sdk';
+import { ProjectMeta, Client, GalleryMeta } from '@valist/sdk';
 import { handleEvent } from './events';
 import * as utils from './utils';
 import { normalizeError, refineYouTube } from './common';
@@ -38,26 +38,31 @@ export const schema = z.object({
   linkRepository: z.boolean(),
 });
 
+const isFile = (file: File | String): file is File => {
+  return (file as File).lastModified !== undefined;
+};
+
 export async function updateProject(
   address: string | undefined,
   projectId: string,
-  image: File | string,
-  mainCapsule: File | string,
-  gallery: File[],
-  repository: string,
+  oldMeta: ProjectMeta,
+  ytLink: string,
+  image: File | undefined,
+  mainCapsule: File | undefined,
+  gallery: (File | String)[],
   values: FormValues,
   valist: Client,
   cache: ApolloCache<any>,
   chainId: number,
-): Promise<boolean | undefined> {
+): Promise<undefined | ProjectMeta> {
   try {
-  	utils.hideError();
-
-    if (!address) {
-      throw new Error('connect your wallet to continue');
-    }
+    utils.hideError();
+    if (!address) throw new Error('connect your wallet to continue');
+    utils.showLoading('Uploading files');
 
     const meta: ProjectMeta = {
+      image: oldMeta.image,
+      main_capsule: oldMeta.main_capsule,
       name: values.displayName,
       short_description: values.shortDescription,
       description: values.description,
@@ -65,41 +70,42 @@ export async function updateProject(
       type: values.type,
       tags: values.tags,
       gallery: [],
-      repository,
+      repository: oldMeta.repository,
       launch_external: values.launchExternal,
       donation_address: values.donationAddress,
       prompt_donation: values.promptDonation,
     };
-
-    utils.showLoading('Uploading files');
-
-    if (typeof image === 'string') {
-      meta.image = image;
-    } else {
+    
+    if (image) {
       meta.image = await valist.writeFile(image, false, (progress: number) => {
         utils.updateLoading(`Uploading ${image.name}: ${progress}%`);
       });
     };
 
-    if (typeof mainCapsule === 'string') {
-      meta.main_capsule = mainCapsule;
-    } else {
+    if (mainCapsule) {
       meta.main_capsule = await valist.writeFile(mainCapsule, false, (progress: number) => {
         utils.updateLoading(`Uploading ${mainCapsule.name}: ${progress}%`);
       });
     };
 
-    if (values.youTubeLink) {
-      const src = values.youTubeLink;
-      meta.gallery?.push({ name: '', type: 'youtube', src });
-    };
-
     for (const item of gallery) {
-      const src = await valist.writeFile(item, false, (progress: number) => {  
-        utils.updateLoading(`Uploading ${item.name}: ${progress}%`);
-      });
-      meta.gallery?.push({ name: '', type: 'image', src });
-    };
+      if (typeof item === 'string') {
+        meta.gallery?.push({ name: '', type: 'image', src: item });
+      } else if (isFile(item)) {
+        const src = await valist.writeFile(item, false, (progress: number) => {  
+          utils.updateLoading(`Uploading ${item.name}: ${progress}%`);
+        });
+        meta.gallery?.push({ name: '', type: 'image', src });
+      }
+    }
+  
+    const imgGallery = meta?.gallery?.filter((item: GalleryMeta) => item.type === 'image') || [];
+    const isNewYt = ytLink !== values.youTubeLink;
+    if ((isNewYt && values.youTubeLink) || (!isNewYt && ytLink)) {
+      meta.gallery = [{ name: '', type: 'youtube', src: values.youTubeLink || ytLink }, ...imgGallery];
+    } else {
+      meta.gallery = imgGallery;
+    }
 
     utils.updateLoading('Creating transaction');
     const transaction = await valist.setProjectMeta(projectId, meta);
@@ -110,13 +116,13 @@ export async function updateProject(
     const receipt = await transaction.wait();
     receipt.events?.forEach(event => handleEvent(event, cache));
 
-    return true;
+    return meta;
   } catch(error: any) {
     utils.showError(normalizeError(error));
   } finally {
     utils.hideLoading();
   };
-};
+}
 
 export async function addProjectMember(
   address: string | undefined,
